Keep the rules engine entry points covered in one place
The new test class re-tested what RulesEngineEvaluateTest already covered,
leaving two files for one object and no rule for where a new case belongs.
The iOS counterpart put its transform tests in the existing evaluate test
file; doing the same here keeps the two suites readable side by side.
